Move NamenodeStats to file system abstraction
