+++ lib/trino-hdfs/src/main/java/io/trino/hdfs/HdfsNamenodeStats.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.hdfs;
+
+import org.weakref.jmx.Managed;
+import org.weakref.jmx.Nested;
+
+public final class HdfsNamenodeStats
+{
+    private final CallStats listLocatedStatus = new CallStats();
+    private final CallStats remoteIteratorNext = new CallStats();
+
+    @Managed
+    @Nested
+    public CallStats getListLocatedStatus()
+    {
+        return listLocatedStatus;
+    }
+
+    @Managed
+    @Nested
+    public CallStats getRemoteIteratorNext()
+    {
+        return remoteIteratorNext;
+    }
+}
