+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/AbstractTestHiveFileSystem.java
@@ -28,7 +28,7 @@ import io.trino.hdfs.HdfsConfig;
 import io.trino.hdfs.HdfsConfiguration;
 import io.trino.hdfs.HdfsContext;
 import io.trino.hdfs.HdfsEnvironment;
-import io.trino.hdfs.NamenodeStats;
+import io.trino.hdfs.HdfsNamenodeStats;
 import io.trino.hdfs.authentication.NoHdfsAuthentication;
 import io.trino.operator.GroupByHashPageIndexerFactory;
 import io.trino.plugin.base.CatalogName;
@@ -231,7 +231,7 @@ public abstract class AbstractTestHiveFileSystem
                 transactionManager,
                 hivePartitionManager,
                 new HdfsFileSystemFactory(hdfsEnvironment, HDFS_FILE_SYSTEM_STATS),
-                new NamenodeStats(),
+                new HdfsNamenodeStats(),
                 hdfsEnvironment,
                 new BoundedExecutor(executor, config.getMaxSplitIteratorThreads()),
                 new CounterStat(),
@@ -474,7 +474,7 @@ public abstract class AbstractTestHiveFileSystem
                 basePath,
                 fs,
                 new FileSystemDirectoryLister(),
-                new NamenodeStats(),
+                new HdfsNamenodeStats(),
                 HiveFileIterator.NestedDirectoryPolicy.RECURSE,
                 false); // ignoreAbsentPartitions
 
@@ -490,7 +490,7 @@ public abstract class AbstractTestHiveFileSystem
                 basePath,
                 fs,
                 new FileSystemDirectoryLister(),
-                new NamenodeStats(),
+                new HdfsNamenodeStats(),
                 HiveFileIterator.NestedDirectoryPolicy.IGNORED,
                 false); // ignoreAbsentPartitions
         List<Path> shallowListing = Streams.stream(shallowIterator)
